requirements.txt
flask-restplus
flask
sqlalchemy
psycopg2
ipython
geojson
pillow
pytest
pyjwt
